Revert "Merge commit '4e62ccd90106' into HEAD"
This reverts commit 15a59da391629943d944b7fce3843103b051a380, reversing
changes made to 1e0775ff71ef04156b2832ae0a8213a06e202d73.

+++ include/linux/mlx4/device.h
@@ -583,7 +583,7 @@ struct mlx4_uar {
 };
 
 struct mlx4_bf {
-	unsigned int		offset;
+	unsigned long		offset;
 	int			buf_size;
 	struct mlx4_uar	       *uar;
 	void __iomem	       *reg;
